feat: add webpack build - FUI-2140
