+++ client-tmp/react/src/utils/index.ts
@@ -2,5 +2,4 @@ export * from './history';
 export * from './fdc3';
 export * from './permissions';
 export * from './setApiHost';
-export * from './store';
 export * from './getLayoutNameByRoute';
