+++ client-tmp/react/webpack.config.js
@@ -92,7 +92,6 @@ module.exports = (env, argv) => {
       compress: true,
       port: 3000,
       historyApiFallback: true,
-      overlay: false,
     },
     stats: {
       all: false,
